elife-backend/requirements.txt
absl-py==2.2.2
alembic==1.15.2
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
easyocr==1.7.2
filelock==3.18.0
firebase-admin==6.8.0
Flask==3.1.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==25.2.10
fonttools==4.58.0
fsspec==2025.3.2
fuzzywuzzy==0.18.0
gast==0.6.0
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.62.3
h5py==3.13.0
httplib2==0.22.0
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
jax==0.6.0
jaxlib==0.6.0
Jinja2==3.1.6
joblib==1.5.0
keras==3.9.2
keras-facenet==0.3.2
kiwisolver==1.4.8
lazy_loader==0.4
Levenshtein==0.27.1
libclang==18.1.1
lz4==4.4.4
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
mtcnn==1.0.0
namex==0.0.9
networkx==3.4.2
ninja==1.11.1.4
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==4.25.7
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.22
Pygments==2.19.1
PyJWT==2.8.0
pyparsing==3.2.3
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.27.1
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.3
rich==14.0.0
rsa==4.9.1
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
sentencepiece==0.2.0
shapely==2.1.0
six==1.17.0
sounddevice==0.5.1
SQLAlchemy==2.0.41
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.10
torch==2.7.0
torchvision==0.22.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
